+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ModelBuilder`: Basic support for copy&paste
 
+### Fixed
+
+- `Collection`: Only show providers heading if there are providers available.
+
 ## [2.4.0] - 2020-08-02
 
 ### Added